Use canister_sig_util instead of canister_sig_util_br.
