+++ Cargo.toml
@@ -1,7 +1,6 @@
 [workspace]
 members = [
     "src/internet_identity",
-    "src/canister_sig_util_br",
     "src/canister_sig_util",
     "src/canister_tests",
     "src/internet_identity_interface",
