+++ Dockerfile
@@ -48,7 +48,6 @@ COPY src/internet_identity_interface/Cargo.toml src/internet_identity_interface/
 COPY src/archive/Cargo.toml src/archive/Cargo.toml
 COPY src/canister_tests/Cargo.toml src/canister_tests/Cargo.toml
 COPY src/canister_sig_util/Cargo.toml src/canister_sig_util/Cargo.toml
-COPY src/canister_sig_util_br/Cargo.toml src/canister_sig_util_br/Cargo.toml
 COPY src/vc_util_br/Cargo.toml src/vc_util_br/Cargo.toml
 ENV CARGO_TARGET_DIR=/cargo_target
 COPY ./scripts/build ./scripts/build
@@ -62,8 +61,6 @@ RUN mkdir -p src/internet_identity/src \
     && touch src/canister_tests/src/lib.rs \
     && mkdir -p src/canister_sig_util/src \
     && touch src/canister_sig_util/src/lib.rs \
-    && mkdir -p src/canister_sig_util_br/src \
-    && touch src/canister_sig_util_br/src/lib.rs \
     && mkdir -p src/vc_util_br/src \
     && touch src/vc_util_br/src/lib.rs \
     && ./scripts/build --only-dependencies --internet-identity --archive \
