+++ demos/vc_issuer/Cargo.toml
@@ -8,7 +8,7 @@ edition = "2021"
 
 [dependencies]
 # local dependencies
-canister_sig_util_br = { path = "../../src/canister_sig_util_br" }
+canister_sig_util = { path = "../../src/canister_sig_util" }
 internet_identity_interface = { path = "../../src/internet_identity_interface" }
 vc_util = { path = "../../src/vc_util_br" }
 # ic dependencies
