--- src/canister_sig_util_br/Cargo.toml
@@ -1,23 +1,0 @@
-[package]
-name = "canister_sig_util_br"
-description = "Utils for handling canister signatures (branch version)"
-version = "0.1.0"
-edition = "2021"
-
-[dependencies]
-# ic dependencies
-candid = "0.9"
-ic-cdk = "0.10"
-ic-certification = "1.2.0"
-ic-certified-map = "0.4"
-hex = "0.4.3"
-
-# other dependencies
-lazy_static = "1.4"
-serde = { version = "1", features = ["rc"] }
-serde_bytes = "0.11"
-serde_cbor = "0.11"
-sha2 = "^0.10" # set bound to match ic-certified-map bound
-
-[dev-dependencies]
-rand = { version ="0.8.5" }
