+++ src/internet_identity/src/vc_mvp.rs
@@ -2,7 +2,7 @@ use crate::assets::CertifiedAssets;
 use crate::delegation::check_frontend_length;
 use crate::{delegation, hash, state, update_root_hash, LABEL_SIG, MINUTE_NS};
 use candid::Principal;
-use canister_sig_util_br::canister_sig_pk_der;
+use canister_sig_util::get_canister_sig_pk_der;
 use ic_cdk::api::{data_certificate, time};
 use ic_cdk::trap;
 use ic_certified_map::{Hash, HashTree};
@@ -45,7 +45,7 @@ pub async fn prepare_id_alias(
     let id_alias_principal = get_id_alias_principal(identity_number, &dapps);
     let seed = calculate_id_alias_seed(identity_number, &dapps);
     let canister_id = ic_cdk::id();
-    let canister_sig_pk_der = canister_sig_pk_der(canister_id, &seed);
+    let canister_sig_pk_der = get_canister_sig_pk_der(canister_id, &seed);
 
     let rp_tuple = AliasTuple {
         id_alias: id_alias_principal,
@@ -99,7 +99,7 @@ pub fn get_id_alias(
         let id_rp = delegation::get_principal(identity_number, dapps.relying_party.clone());
         let id_issuer = delegation::get_principal(identity_number, dapps.issuer.clone());
         let canister_id = ic_cdk::id();
-        let canister_sig_pk_der = canister_sig_pk_der(canister_id, &seed);
+        let canister_sig_pk_der = get_canister_sig_pk_der(canister_id, &seed);
 
         let signing_input = vc_signing_input(rp_id_alias_jwt, &canister_sig_pk_der, canister_id);
         let msg_hash = vc_signing_input_hash(&signing_input);
