+++ src/vc_util_br/Cargo.toml
@@ -10,7 +10,7 @@ candid = "0.9"
 ic-certified-map = "0.4"
 ic-crypto-standalone-sig-verifier = { git = "https://github.com/dfinity/ic", rev = "bf9bc00b032a18d2dbcfcfcfb7a76a562f350c9e" }
 ic-types = { git = "https://github.com/dfinity/ic", rev = "bf9bc00b032a18d2dbcfcfcfb7a76a562f350c9e" }
-canister_sig_util_br = { path = "../canister_sig_util_br" }
+canister_sig_util = { path = "../canister_sig_util" }
 
 # vc dependencies
 identity_core = { git = "https://github.com/frederikrothenberger/identity.rs.git", branch = "frederik/wasm-test", default-features = false }
